requirements.txt
# Geospatial & GIS
geopandas>=1.0.0
owslib>=0.30.0
pyproj>=3.6.0
rasterio>=1.3.0
shapely>=2.0.0
fiona>=1.9.0

# Computer Vision & Deep Learning
tensorflow>=2.15.0
segmentation-models>=1.2.1
opencv-python>=4.8.0
pillow>=10.0.0

# Data Analysis & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
